// src/app/models/user-management/autRoleData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AllMenuData } from './allMenuData';
import { AutUserData } from 'src/app/models/user-management';

/**
 * custom สิทธิ์
 */
export interface AutRoleData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** การอนุญาต */
    allMenu?: AllMenuData[];
    first?: number;
    mode?: string;
    /** ชื่อสิทธิ์ */
    name?: string;
    /** PK */
    roleId?: number;
    rowNum?: number;
    size?: number;

    /** custom user role */
    userList?: AutUserData[];

}