// src/app/models/user-management/autUserData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AutRoleData } from './autRoleData';

export interface AutUserData {
    /** ระดับการเข้าถึง */
    accessLevel?: number;
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** วันที่ให้ดำเนินการสิ้นสุด */
    activePeriodEnd?: string;
    /** วันที่ให้ดำเนินการเริ่มต้น */
    activePeriodStart?: string;
    /** กำหนดวันที่ให้ดำเนินการ */
    activePeriodStatus?: boolean;
    /** count */
    count?: number;
    /** custom รหัสคณะ/หน่วยงาน */
    depCode?: string;
    /** คณะ (mas_department) */
    depIdLevel1?: number;
    /** ภาควิชา (mas_department) */
    depIdLevel2?: number;
    /** custom ตัวย่อคณะ/หน่วยงาน (ภาษาอังกฤษ) */
    depNameAbbrEn?: string;
    /** custom ตัวย่อคณะ/หน่วยงาน (ภาษาไทย) */
    depNameAbbrTh?: string;
    /** custom ชื่อเต็มคณะ/หน่วยงาน (ภาษาอังกฤษ) */
    depNameEn?: string;
    /** custom ชื่อย่อคณะ/หน่วยงาน (ภาษาอังกฤษ) */
    depNameShortEn?: string;
    /** custom ชื่อย่อคณะ/หน่วยงาน (ภาษาไทย) */
    depNameShortTh?: string;
    /** custom ชื่อเต็มคณะ/หน่วยงาน (ภาษาไทย) */
    depNameTh?: string;
    /** อีเมล */
    email?: string;
    /** FileName */
    filename?: string;
    first?: number;
    /** ชื่อ (อังกฤษ) */
    firstnameEn?: string;
    /** ชื่อ (ไทย) */
    firstnameTh?: string;
    /** custom แยก route by condition */
    isExists?: boolean;
    /** Login ผ่าน Lifelong */
    isLocal?: boolean;
    /** วันที่เปลี่ยนรหัสผ่านล่าสุด */
    lastChangepasswdDatetime?: string;
    /** วันที่เข้าสู่ระบบล่าสุด */
    lastLoginDatetime?: string;
    /** วันที่ออกจากระบบล่าสุด */
    lastLogoutDatetime?: string;
    /** นามสกุล (อังกฤษ) */
    lastnameEn?: string;
    /** นามสกุล (ไทย) */
    lastnameTh?: string;
    mode?: string;
    /** Module */
    module?: number;
    /** รหัสผ่าน */
    password?: string;
    /** Personal Id */
    personalId?: number;
    /** Prefix */
    prefix?: string;
    /** รหัสอ้างอิง ID ผู้ใช้งาน */
    refUserId?: number;
    /** ประเภทข้อมูลอ้างอิงผู้ใช้งาน 30032000  */
    refUserType?: number;
    /** custom pkId สิทธิ์ */
    roleId?: number;
    /** custom pkIds สิทธิ์ */
    roleIds?: number[];
    /** custom สิทธิ์ */
    roles?: AutRoleData[];
    rowNum?: number;
    size?: number;
    /** เบอร์โทรศัพท์ */
    telephone?: string;
    /** PK */
    userId?: number;
    /** ชื่อผู้ใช้งาน */
    username?: string;
    /** รูปภาพลายเซ็นต์ */
    userSignaturePath?: string;
}